# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HT Lotto API",
        "description": "REST API for HT ភ្នាក់ Lottery Agent App. Roles: admin | master | staff.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/account": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Today's sales totals — staff: own; master: own staff; admin: all",
                "operationId": "f020604f40c86daafc253792926cdb39",
                "responses": {
                    "200": {
                        "description": "Sales summary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "today_sales": {
                                                    "type": "number",
                                                    "example": 150000
                                                },
                                                "today_win": {
                                                    "type": "number",
                                                    "example": 20000
                                                },
                                                "today_count": {
                                                    "type": "integer",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/account/sales/detail": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Today's sales breakdown by session — staff: own; master: own staff; admin: all",
                "operationId": "89308aed350eefcb9f8971d0f7dd0b9f",
                "responses": {
                    "200": {
                        "description": "Sales breakdown",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "session": {
                                                        "type": "string",
                                                        "enum": [
                                                            "morning",
                                                            "noon",
                                                            "evening"
                                                        ]
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "example": 50000
                                                    },
                                                    "win": {
                                                        "type": "number",
                                                        "example": 5000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/account/password": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Change own password — all roles",
                "operationId": "09b5ee785db2535f40e0884cee7c6c0d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "example": "oldpass"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpass123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpass123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong current password or validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}